requirements.txt
pandas>=2.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-experimental>=0.0.47
python-dotenv>=1.0.0
streamlit>=1.30.0
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0
